refactor: FakeRepository를 StateHolder 테스트 파일에서 분리
