+++ app/src/test/java/woowacourse/shopping/presentation/shopping/FakeProductRepository.kt
@@ -1,0 +1,41 @@
+package woowacourse.shopping.presentation.shopping
+
+import woowacourse.shopping.data.ProductFixture
+import woowacourse.shopping.domain.model.product.Product
+import woowacourse.shopping.domain.model.product.Products
+import woowacourse.shopping.domain.repository.ProductRepository
+import kotlin.math.min
+import kotlin.uuid.ExperimentalUuidApi
+import kotlin.uuid.Uuid
+
+class FakeProductRepository(
+    private val products: Products,
+) : ProductRepository {
+    override fun getProducts(): Products = products
+
+    override fun getPagingProducts(
+        page: Int,
+        pageSize: Int,
+    ): Products {
+        if (page < 0 || pageSize <= 0) return Products()
+
+        val fromIndex = page * pageSize
+
+        if (fromIndex >= products.productItems.size) {
+            return Products()
+        }
+
+        val toIndex = min(fromIndex + pageSize, products.productItems.size)
+        return Products(products.productItems.subList(fromIndex, toIndex))
+    }
+
+    override fun hasNextPage(
+        currentPage: Int,
+        pageSize: Int,
+    ): Boolean =
+        getPagingProducts(page = currentPage, pageSize = pageSize).productItems.size >= pageSize &&
+            (currentPage + 1) * pageSize < products.productItems.size
+
+    @OptIn(ExperimentalUuidApi::class)
+    override fun findProductById(productId: Uuid): Product? = ProductFixture.productList.firstOrNull { it.productId == productId }
+}
